html/semantics/forms/the-label-element/labelable-elements.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: labelable elements</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form style="display:none">
  <output id="testoutput"></output>
  <label id="lbl0" for="testoutput"></label>
  <progress id="testprogress"></progress>
  <label id="lbl1" for="testprogress"></label>
  <select id="testselect"></select>
  <label id="lbl2" for="testselect"></label>
  <textarea id="testarea"></textarea>
  <label id="lbl3" for="testarea"></label>
  <button id="testButton"></button>
  <label id="lbl4" for="testButton"></label>
  <input type="hidden" id="testHidden">
  <label id="lbl5" for="testHidden"></label>
  <input type="radio" id="testRadio">
  <label id="lbl6" for="testRadio"></label>
  <keygen id="testkeygen"></keygen>
  <label id="lbl7" for="testkeygen"></label>
  <meter id="testmeter"></meter>
  <label id="lbl8" for="testmeter"></label>
  <label id="lbl9">
    <input type="radio" id="Radio1">
    <input type="radio" id="Radio2">
  </label>
</form>

<script>

test(function() {
  assert_equals(document.getElementById("lbl0").control.id, "testoutput", "The labeled control should be an output.");
}, "Check if the output element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl1").control.id, "testprogress", "The labeled control should be a progress.");
}, "Check if the progress element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl2").control.id, "testselect", "The labeled control should be a select.");
}, "Check if the select element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl3").control.id, "testarea", "The labeled control should be a textarea.");
}, "Check if the textarea element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl4").control.id, "testButton", "The labeled control should be a button.");
}, "Check if the button element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl5").control, null, "The labeled control should not be an input in hidden state.");
}, "Check if the hidden input element is not a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl6").control.id, "testRadio", "The labeled control should be an input in radio state.");
}, "Check if the input element in radio state is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl7").control.id, "testkeygen", "The labeled control should be a keygen.");
}, "Check if the keygen element is a labelable form-associated element");

test(function() {
  assert_equals(document.getElementById("lbl8").control.id, "testmeter", "The labeled control should be a meter.");
}, "Check if the meter element is a labelable form-associated element");

</script>
